front-end PR comments; duration snap tests, rhs_list_incident component
