+++ webapp/package.json
@@ -37,6 +37,7 @@
     "@types/react-dom": "16.9.8",
     "@types/react-intl": "3.0.0",
     "@types/react-router-dom": "5.1.5",
+    "@types/react-test-renderer": "^16.9.3",
     "@types/react-transition-group": "4.4.0",
     "@types/styled-components": "^5.1.1",
     "@typescript-eslint/eslint-plugin": "3.1.0",
@@ -65,6 +66,7 @@
     "node-sass": "4.14.1",
     "react-beautiful-dnd": "13.0.0",
     "react-bootstrap": "github:mattermost/react-bootstrap#c6957962364e0818a51bbfd13e89919903b422d6",
+    "react-test-renderer": "^17.0.1",
     "sass-loader": "8.0.2",
     "style-loader": "1.2.1",
     "webpack": "4.43.0",
