+++ webapp/src/components/rhs/rhs_list_view.tsx
@@ -4,19 +4,16 @@
 import React from 'react';
 import {useDispatch, useSelector} from 'react-redux';
 import Scrollbars from 'react-custom-scrollbars';
-import styled, {css} from 'styled-components';
+import styled from 'styled-components';
 
 import {getCurrentTeam} from 'mattermost-redux/selectors/entities/teams';
 import {GlobalState} from 'mattermost-redux/types/store';
 import {Team} from 'mattermost-redux/types/teams';
 import {getCurrentChannelId} from 'mattermost-redux/selectors/entities/channels';
-import {getPost} from 'mattermost-redux/selectors/entities/posts';
-import {Post} from 'mattermost-redux/types/posts';
 
 import {pluginId} from 'src/manifest';
 import RHSWelcomeView from 'src/components/rhs/rhs_welcome_view';
 import PlusIcon from 'src/components/assets/icons/plus_icon';
-import Profile from 'src/components/profile/profile';
 import {
     renderThumbVertical,
     renderTrackHorizontal,
@@ -25,10 +22,10 @@ import {
 import {setRHSViewingIncident, startIncident} from 'src/actions';
 import {navigateToTeamPluginUrl, navigateToUrl} from 'src/browser_routing';
 import {Incident} from 'src/types/incident';
-import Duration from 'src/components/duration';
 import DotMenu, {DropdownMenuItem} from 'src/components/dot_menu';
 import {myActiveIncidentsList} from 'src/selectors';
 import {HamburgerButton} from 'src/components/assets/icons/three_dots_icon';
+import RHSListIncident from 'src/components/rhs/rhs_list_incident';
 
 const Header = styled.div`
     display: grid;
@@ -63,75 +60,6 @@ const Link = styled.span`
     }
 `;
 
-interface IncidentContainerProps {
-    active: boolean;
-}
-
-const IncidentContainer = styled.div<IncidentContainerProps>`
-    display: flex;
-    flex-direction: column;
-    padding: 20px;
-    box-shadow: inset 0px -1px 0px var(--center-channel-color-24);
-
-    ${(props) => props.active && css`
-        box-shadow: inset 0px -1px 0px var(--center-channel-color-24), inset 4px 0px 0px var(--button-bg);
-    `}
-`;
-
-const IncidentTitle = styled.div`
-    font-size: 14px;
-    font-style: normal;
-    font-weight: 600;
-    line-height: 20px;
-    letter-spacing: 0;
-    text-align: left;
-    margin-bottom: 6px;
-`;
-
-const Row = styled.div`
-    display: flex;
-    flex-direction: row;
-    flex-wrap: nowrap;
-    font-size: 12px;
-    line-height: 20px;
-    margin: 1px 0;
-`;
-
-const Col1 = styled.div`
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-    font-weight: 600;
-`;
-
-const Col2 = styled.div`
-    display: flex;
-    flex-direction: column;
-    flex: 3;
-    font-weight: 400;
-`;
-
-const SmallerProfile = styled(Profile)`
-    >.image {
-        width: 20px;
-        height: 20px;
-    }
-`;
-
-const Button = styled.button`
-    display: block;
-    border: 1px solid var(--button-bg);
-    border-radius: 4px;
-    background: transparent;
-    font-size: 12px;
-    font-weight: 600;
-    line-height: 9.5px;
-    color: var(--button-bg);
-    text-align: center;
-    padding: 10px 0;
-    margin-top: 10px;
-`;
-
 const Footer = styled.div`
     display: block;
     font-size: 12px;
@@ -148,8 +76,6 @@ const RHSListView = () => {
     const currentTeam = useSelector<GlobalState, Team>(getCurrentTeam);
     const currentChannelId = useSelector<GlobalState, string>(getCurrentChannelId);
     const incidentList = useSelector<GlobalState, Incident[]>(myActiveIncidentsList);
-    type GetPostType = (postId: string) => Post;
-    const getPostFromState = useSelector<GlobalState, GetPostType>((state) => (postId) => getPost(state, postId));
 
     const viewIncident = (channelId: string) => {
         dispatch(setRHSViewingIncident());
@@ -191,50 +117,13 @@ const RHSListView = () => {
                     </Header>
 
                     {incidentList.map((incident) => {
-                        const lastUpdated = findLastUpdated(getPostFromState, incident.status_posts_ids);
-
                         return (
-                            <IncidentContainer
+                            <RHSListIncident
                                 key={incident.id}
+                                incident={incident}
                                 active={currentChannelId === incident.channel_id}
-                            >
-                                <IncidentTitle>{incident.name}</IncidentTitle>
-                                <Row>
-                                    <Col1>{'Stage:'}</Col1>
-                                    <Col2>{incident.active_stage_title}</Col2>
-                                </Row>
-                                <Row>
-                                    <Col1>{'Duration:'}</Col1>
-                                    <Col2>
-                                        <Duration
-                                            from={incident.create_at}
-                                            to={incident.end_at}
-                                        />
-                                    </Col2>
-                                </Row>
-                                <Row>
-                                    <Col1>{'Last updated:'}</Col1>
-                                    <Col2>
-                                        <Duration
-                                            from={lastUpdated}
-                                            to={0}
-                                            ago={true}
-                                        />
-                                    </Col2>
-                                </Row>
-                                <Row>
-                                    <Col1>{'Commander:'}</Col1>
-                                    <Col2>
-                                        <SmallerProfile userId={incident.commander_user_id}/>
-                                    </Col2>
-                                </Row>
-                                <Button
-                                    onClick={() => viewIncident(incident.channel_id)}
-                                    data-testid='go-to-channel'
-                                >
-                                    {'Go to Incident Channel'}
-                                </Button>
-                            </IncidentContainer>
+                                viewIncident={viewIncident}
+                            />
                         );
                     })}
 
@@ -269,18 +158,5 @@ const ThreeDotMenu = (props: ThreeDotMenuProps) => (
         />
     </DotMenu>
 );
-
-const findLastUpdated = (getPostFromState: (postId: string) => Post, updatePostIDs: string[]): number => {
-    const postIDs = [...updatePostIDs].reverse();
-
-    for (const postId of postIDs) {
-        const post = getPostFromState(postId);
-        if (post?.delete_at === 0) {
-            return post.create_at;
-        }
-    }
-
-    return 0;
-};
 
 export default RHSListView;
